Refactor test setup into its own file
